Add Global, Missing and Multi Terms aggs
